bescribe: Touch all the things in the name of bescribe.
Enacts broad, sweeping changes to the way the Page obejct used
in each `it` block is constructed and provided. This is facilitated
by the new `bescribe` helper, which also sets up before and after
runners for the given test script.

Also, be refactoring **when** the session is created, the issue with
all browsers being created immediately has been resolved.

bescribe is synonymous is mocha's `describe` and even has support for
`describe.only` through `bescribe.only`.